replantear grraphs

+++ app/src/main/java/com/fup/accoffe/ui/drying/DryingListFragment.kt
@@ -1,5 +1,7 @@
 package com.fup.accoffe.ui.drying
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
 import android.util.Log
@@ -101,15 +103,33 @@ class DryingListFragment : Fragment() {
                 activity?.runOnUiThread {
                     val adapter = DryingListAdapter(dataList, onClickDelete = {
 
-                        db.collection(collectionName).document(it).delete()
-                            .addOnSuccessListener {
-                                Toast.makeText(context, "Eliminado Correctamente", Toast.LENGTH_SHORT).show()
-                                fetchAllDataFromFirestore(collectionName)
-                            }
-                            .addOnFailureListener { e ->
-                                Toast.makeText(context, "Error al Eliminar", Toast.LENGTH_SHORT).show()
+                        val builder = AlertDialog.Builder(requireContext())
 
-                            }
+                        builder.setTitle("Confirmación")
+                        builder.setMessage("¿Estás seguro de que deseas borrar?")
+
+                        builder.setPositiveButton("Sí") { dialogInterface: DialogInterface, i: Int ->
+
+
+                            db.collection(collectionName).document(it).delete()
+                                .addOnSuccessListener {
+                                    Toast.makeText(context, "Eliminado Correctamente", Toast.LENGTH_SHORT).show()
+                                    fetchAllDataFromFirestore(collectionName)
+                                }
+                                .addOnFailureListener { e ->
+                                    Toast.makeText(context, "Error al Eliminar", Toast.LENGTH_SHORT).show()
+
+                                }
+                        }
+
+                        builder.setNegativeButton("No") { dialogInterface: DialogInterface, i: Int ->
+
+                        }
+
+                        val dialog = builder.create()
+                        dialog.show()
+
+
 
                     },onClickInfo = {
                         val bundle = Bundle()
